Added Dockerfile and docker-compose
I actually have dealt with dockerfiles and docker-compose before when once setting up a mediawiki installation on a raspberry pi once. Once I realized that's what this process was, I gave myself a refresher, created the simple Dockerfile to contain everything, then made the docker-compose file to run the code inside of the Docker.